Move remove-related tests to snapstate_remove_test.go
