+++ overlord/snapstate/snapstate_test.go
@@ -42,7 +42,6 @@ import (
 	"github.com/snapcore/snapd/gadget"
 	"github.com/snapcore/snapd/interfaces"
 	"github.com/snapcore/snapd/logger"
-	"github.com/snapcore/snapd/osutil"
 	"github.com/snapcore/snapd/overlord"
 	"github.com/snapcore/snapd/overlord/auth"
 	"github.com/snapcore/snapd/overlord/configstate/config"
@@ -959,196 +958,6 @@ func (s *snapmgrTestSuite) TestDoUpdateHadSlots(c *C) {
 	c.Check(snapsup.PlugsOnly, Equals, false)
 }
 
-func (s *snapmgrTestSuite) TestRemoveTasks(c *C) {
-	s.state.Lock()
-	defer s.state.Unlock()
-
-	snapstate.Set(s.state, "foo", &snapstate.SnapState{
-		Active: true,
-		Sequence: []*snap.SideInfo{
-			{RealName: "foo", Revision: snap.R(11)},
-		},
-		Current:  snap.R(11),
-		SnapType: "app",
-	})
-
-	ts, err := snapstate.Remove(s.state, "foo", snap.R(0), nil)
-	c.Assert(err, IsNil)
-
-	c.Assert(s.state.TaskCount(), Equals, len(ts.Tasks()))
-	verifyRemoveTasks(c, ts)
-}
-
-func (s *snapmgrTestSuite) TestRemoveTasksAutoSnapshotDisabled(c *C) {
-	snapstate.AutomaticSnapshot = func(st *state.State, instanceName string) (ts *state.TaskSet, err error) {
-		return nil, snapstate.ErrNothingToDo
-	}
-
-	s.state.Lock()
-	defer s.state.Unlock()
-
-	snapstate.Set(s.state, "foo", &snapstate.SnapState{
-		Active: true,
-		Sequence: []*snap.SideInfo{
-			{RealName: "foo", Revision: snap.R(11)},
-		},
-		Current:  snap.R(11),
-		SnapType: "app",
-	})
-
-	ts, err := snapstate.Remove(s.state, "foo", snap.R(0), nil)
-	c.Assert(err, IsNil)
-
-	c.Assert(taskKinds(ts.Tasks()), DeepEquals, []string{
-		"stop-snap-services",
-		"run-hook[remove]",
-		"auto-disconnect",
-		"remove-aliases",
-		"unlink-snap",
-		"remove-profiles",
-		"clear-snap",
-		"discard-snap",
-	})
-}
-
-func (s *snapmgrTestSuite) TestRemoveTasksAutoSnapshotDisabledByPurgeFlag(c *C) {
-	s.state.Lock()
-	defer s.state.Unlock()
-
-	snapstate.Set(s.state, "foo", &snapstate.SnapState{
-		Active: true,
-		Sequence: []*snap.SideInfo{
-			{RealName: "foo", Revision: snap.R(11)},
-		},
-		Current:  snap.R(11),
-		SnapType: "app",
-	})
-
-	ts, err := snapstate.Remove(s.state, "foo", snap.R(0), &snapstate.RemoveFlags{Purge: true})
-	c.Assert(err, IsNil)
-
-	c.Assert(taskKinds(ts.Tasks()), DeepEquals, []string{
-		"stop-snap-services",
-		"run-hook[remove]",
-		"auto-disconnect",
-		"remove-aliases",
-		"unlink-snap",
-		"remove-profiles",
-		"clear-snap",
-		"discard-snap",
-	})
-}
-
-func (s *snapmgrTestSuite) TestRemoveHookNotExecutedIfNotLastRevison(c *C) {
-	s.state.Lock()
-	defer s.state.Unlock()
-
-	snapstate.Set(s.state, "foo", &snapstate.SnapState{
-		Active: true,
-		Sequence: []*snap.SideInfo{
-			{RealName: "foo", Revision: snap.R(11)},
-			{RealName: "foo", Revision: snap.R(12)},
-		},
-		Current: snap.R(12),
-	})
-
-	ts, err := snapstate.Remove(s.state, "foo", snap.R(11), nil)
-	c.Assert(err, IsNil)
-
-	runHooks := tasksWithKind(ts, "run-hook")
-	// no 'remove' hook task
-	c.Assert(runHooks, HasLen, 0)
-}
-
-func (s *snapmgrTestSuite) TestRemoveConflict(c *C) {
-	s.state.Lock()
-	defer s.state.Unlock()
-
-	snapstate.Set(s.state, "some-snap", &snapstate.SnapState{
-		Active:   true,
-		Sequence: []*snap.SideInfo{{RealName: "some-snap", Revision: snap.R(11)}},
-		Current:  snap.R(11),
-	})
-
-	ts, err := snapstate.Remove(s.state, "some-snap", snap.R(0), nil)
-	c.Assert(err, IsNil)
-	// need a change to make the tasks visible
-	s.state.NewChange("remove", "...").AddAll(ts)
-
-	_, err = snapstate.Remove(s.state, "some-snap", snap.R(0), nil)
-	c.Assert(err, ErrorMatches, `snap "some-snap" has "remove" change in progress`)
-}
-
-func (s *snapmgrTestSuite) testRemoveDiskSpaceCheck(c *C, featureFlag, automaticSnapshot bool) error {
-	s.state.Lock()
-	defer s.state.Unlock()
-
-	restore := snapstate.MockOsutilCheckFreeSpace(func(string, uint64) error {
-		// osutil.CheckFreeSpace shouldn't be hit if either featureFlag
-		// or automaticSnapshot is false. If both are true then we return disk
-		// space error which should result in snapstate.InsufficientSpaceError
-		// on remove().
-		return &osutil.NotEnoughDiskSpaceError{}
-	})
-	defer restore()
-
-	var automaticSnapshotCalled bool
-	snapstate.AutomaticSnapshot = func(st *state.State, instanceName string) (ts *state.TaskSet, err error) {
-		automaticSnapshotCalled = true
-		if automaticSnapshot {
-			t := s.state.NewTask("foo", "")
-			ts = state.NewTaskSet(t)
-			return ts, nil
-		}
-		// ErrNothingToDo is returned if automatic snapshots are disabled
-		return nil, snapstate.ErrNothingToDo
-	}
-
-	tr := config.NewTransaction(s.state)
-	tr.Set("core", "experimental.check-disk-space-remove", featureFlag)
-	tr.Commit()
-
-	snapstate.Set(s.state, "some-snap", &snapstate.SnapState{
-		Active:   true,
-		Sequence: []*snap.SideInfo{{RealName: "some-snap", Revision: snap.R(11)}},
-		Current:  snap.R(11),
-		SnapType: "app",
-	})
-
-	_, err := snapstate.Remove(s.state, "some-snap", snap.R(0), nil)
-	c.Assert(automaticSnapshotCalled, Equals, true)
-	return err
-}
-
-func (s *snapmgrTestSuite) TestRemoveDiskSpaceCheckDoesNothingWhenNoSnapshot(c *C) {
-	featureFlag := true
-	snapshot := false
-	err := s.testRemoveDiskSpaceCheck(c, featureFlag, snapshot)
-	c.Assert(err, IsNil)
-}
-
-func (s *snapmgrTestSuite) TestRemoveDiskSpaceCheckDisabledByFeatureFlag(c *C) {
-	featureFlag := false
-	snapshot := true
-	err := s.testRemoveDiskSpaceCheck(c, featureFlag, snapshot)
-	c.Assert(err, IsNil)
-}
-
-func (s *snapmgrTestSuite) TestRemoveDiskSpaceForSnapshotError(c *C) {
-	featureFlag := true
-	snapshot := true
-	// both the snapshot and disk check feature are enabled, so we should hit
-	// the disk check (which fails).
-	err := s.testRemoveDiskSpaceCheck(c, featureFlag, snapshot)
-	c.Assert(err, NotNil)
-
-	diskSpaceErr := err.(*snapstate.InsufficientSpaceError)
-	c.Assert(diskSpaceErr, ErrorMatches, `cannot create automatic snapshot when removing last revision of the snap: insufficient space.*`)
-	c.Check(diskSpaceErr.Path, Equals, filepath.Join(dirs.GlobalRootDir, "/var/lib/snapd"))
-	c.Check(diskSpaceErr.Snaps, DeepEquals, []string{"some-snap"})
-	c.Check(diskSpaceErr.ChangeKind, Equals, "remove")
-}
-
 func (s *snapmgrTestSuite) TestDisableSnapDisabledServicesSaved(c *C) {
 	s.state.Lock()
 	defer s.state.Unlock()
@@ -1455,936 +1264,6 @@ func (s *snapmgrTestSuite) TestEnableSnapMissingDisabledServicesSaved(c *C) {
 
 func makeTestSnap(c *C, snapYamlContent string) (snapFilePath string) {
 	return snaptest.MakeTestSnapWithFiles(c, snapYamlContent, nil)
-}
-
-func (s *snapmgrTestSuite) TestRemoveRunThrough(c *C) {
-	c.Assert(snapstate.KeepAuxStoreInfo("some-snap-id", nil), IsNil)
-	c.Check(snapstate.AuxStoreInfoFilename("some-snap-id"), testutil.FilePresent)
-	si := snap.SideInfo{
-		SnapID:   "some-snap-id",
-		RealName: "some-snap",
-		Revision: snap.R(7),
-	}
-
-	s.state.Lock()
-	defer s.state.Unlock()
-
-	snapstate.Set(s.state, "some-snap", &snapstate.SnapState{
-		Active:   true,
-		Sequence: []*snap.SideInfo{&si},
-		Current:  si.Revision,
-		SnapType: "app",
-	})
-
-	chg := s.state.NewChange("remove", "remove a snap")
-	ts, err := snapstate.Remove(s.state, "some-snap", snap.R(0), nil)
-	c.Assert(err, IsNil)
-	chg.AddAll(ts)
-
-	s.state.Unlock()
-	defer s.se.Stop()
-	s.settle(c)
-	s.state.Lock()
-
-	expected := fakeOps{
-		{
-			op:    "auto-disconnect:Doing",
-			name:  "some-snap",
-			revno: snap.R(7),
-		},
-		{
-			op:   "remove-snap-aliases",
-			name: "some-snap",
-		},
-		{
-			op:   "unlink-snap",
-			path: filepath.Join(dirs.SnapMountDir, "some-snap/7"),
-		},
-		{
-			op:    "remove-profiles:Doing",
-			name:  "some-snap",
-			revno: snap.R(7),
-		},
-		{
-			op:   "remove-snap-data",
-			path: filepath.Join(dirs.SnapMountDir, "some-snap/7"),
-		},
-		{
-			op:   "remove-snap-common-data",
-			path: filepath.Join(dirs.SnapMountDir, "some-snap/7"),
-		},
-		{
-			op:   "remove-snap-data-dir",
-			name: "some-snap",
-			path: filepath.Join(dirs.SnapDataDir, "some-snap"),
-		},
-		{
-			op:    "remove-snap-files",
-			path:  filepath.Join(dirs.SnapMountDir, "some-snap/7"),
-			stype: "app",
-		},
-		{
-			op:   "discard-namespace",
-			name: "some-snap",
-		},
-		{
-			op:   "remove-snap-dir",
-			name: "some-snap",
-			path: filepath.Join(dirs.SnapMountDir, "some-snap"),
-		},
-	}
-	// start with an easier-to-read error if this fails:
-	c.Check(len(s.fakeBackend.ops), Equals, len(expected))
-	c.Assert(s.fakeBackend.ops.Ops(), DeepEquals, expected.Ops())
-	c.Check(s.fakeBackend.ops, DeepEquals, expected)
-
-	// verify snapSetup info
-	tasks := ts.Tasks()
-	for _, t := range tasks {
-		if t.Kind() == "run-hook" {
-			continue
-		}
-		if t.Kind() == "save-snapshot" {
-			continue
-		}
-		snapsup, err := snapstate.TaskSnapSetup(t)
-		c.Assert(err, IsNil)
-
-		var expSnapSetup *snapstate.SnapSetup
-		switch t.Kind() {
-		case "discard-conns":
-			expSnapSetup = &snapstate.SnapSetup{
-				SideInfo: &snap.SideInfo{
-					RealName: "some-snap",
-				},
-			}
-		case "clear-snap", "discard-snap":
-			expSnapSetup = &snapstate.SnapSetup{
-				SideInfo: &snap.SideInfo{
-					RealName: "some-snap",
-					SnapID:   "some-snap-id",
-					Revision: snap.R(7),
-				},
-			}
-		default:
-			expSnapSetup = &snapstate.SnapSetup{
-				SideInfo: &snap.SideInfo{
-					RealName: "some-snap",
-					SnapID:   "some-snap-id",
-					Revision: snap.R(7),
-				},
-				Type:      snap.TypeApp,
-				PlugsOnly: true,
-			}
-
-		}
-
-		c.Check(snapsup, DeepEquals, expSnapSetup, Commentf(t.Kind()))
-	}
-
-	// verify snaps in the system state
-	var snapst snapstate.SnapState
-	err = snapstate.Get(s.state, "some-snap", &snapst)
-	c.Assert(err, Equals, state.ErrNoState)
-	c.Check(snapstate.AuxStoreInfoFilename("some-snap-id"), testutil.FileAbsent)
-
-}
-
-func (s *snapmgrTestSuite) TestParallelInstanceRemoveRunThrough(c *C) {
-	si := snap.SideInfo{
-		RealName: "some-snap",
-		Revision: snap.R(7),
-	}
-
-	s.state.Lock()
-	defer s.state.Unlock()
-
-	// pretend we have both a regular snap and a parallel instance
-	snapstate.Set(s.state, "some-snap_instance", &snapstate.SnapState{
-		Active:      true,
-		Sequence:    []*snap.SideInfo{&si},
-		Current:     si.Revision,
-		SnapType:    "app",
-		InstanceKey: "instance",
-	})
-	snapstate.Set(s.state, "some-snap", &snapstate.SnapState{
-		Active:   true,
-		Sequence: []*snap.SideInfo{&si},
-		Current:  si.Revision,
-		SnapType: "app",
-	})
-
-	chg := s.state.NewChange("remove", "remove a snap")
-	ts, err := snapstate.Remove(s.state, "some-snap_instance", snap.R(0), nil)
-	c.Assert(err, IsNil)
-	chg.AddAll(ts)
-
-	s.state.Unlock()
-	s.settle(c)
-	s.state.Lock()
-
-	expected := fakeOps{
-		{
-			op:    "auto-disconnect:Doing",
-			name:  "some-snap_instance",
-			revno: snap.R(7),
-		},
-		{
-			op:   "remove-snap-aliases",
-			name: "some-snap_instance",
-		},
-		{
-			op:   "unlink-snap",
-			path: filepath.Join(dirs.SnapMountDir, "some-snap_instance/7"),
-		},
-		{
-			op:    "remove-profiles:Doing",
-			name:  "some-snap_instance",
-			revno: snap.R(7),
-		},
-		{
-			op:   "remove-snap-data",
-			path: filepath.Join(dirs.SnapMountDir, "some-snap_instance/7"),
-		},
-		{
-			op:   "remove-snap-common-data",
-			path: filepath.Join(dirs.SnapMountDir, "some-snap_instance/7"),
-		},
-		{
-			op:             "remove-snap-data-dir",
-			name:           "some-snap_instance",
-			path:           filepath.Join(dirs.SnapDataDir, "some-snap"),
-			otherInstances: true,
-		},
-		{
-			op:    "remove-snap-files",
-			path:  filepath.Join(dirs.SnapMountDir, "some-snap_instance/7"),
-			stype: "app",
-		},
-		{
-			op:   "discard-namespace",
-			name: "some-snap_instance",
-		},
-		{
-			op:             "remove-snap-dir",
-			name:           "some-snap_instance",
-			path:           filepath.Join(dirs.SnapMountDir, "some-snap"),
-			otherInstances: true,
-		},
-	}
-	// start with an easier-to-read error if this fails:
-	c.Check(len(s.fakeBackend.ops), Equals, len(expected))
-	c.Assert(s.fakeBackend.ops.Ops(), DeepEquals, expected.Ops())
-	c.Check(s.fakeBackend.ops, DeepEquals, expected)
-
-	// verify snapSetup info
-	tasks := ts.Tasks()
-	for _, t := range tasks {
-		if t.Kind() == "run-hook" {
-			continue
-		}
-		if t.Kind() == "save-snapshot" {
-			continue
-		}
-		snapsup, err := snapstate.TaskSnapSetup(t)
-		c.Assert(err, IsNil)
-
-		var expSnapSetup *snapstate.SnapSetup
-		switch t.Kind() {
-		case "discard-conns":
-			expSnapSetup = &snapstate.SnapSetup{
-				SideInfo: &snap.SideInfo{
-					RealName: "some-snap",
-				},
-				InstanceKey: "instance",
-			}
-		case "clear-snap", "discard-snap":
-			expSnapSetup = &snapstate.SnapSetup{
-				SideInfo: &snap.SideInfo{
-					RealName: "some-snap",
-					Revision: snap.R(7),
-				},
-				InstanceKey: "instance",
-			}
-		default:
-			expSnapSetup = &snapstate.SnapSetup{
-				SideInfo: &snap.SideInfo{
-					RealName: "some-snap",
-					Revision: snap.R(7),
-				},
-				Type:        snap.TypeApp,
-				PlugsOnly:   true,
-				InstanceKey: "instance",
-			}
-
-		}
-
-		c.Check(snapsup, DeepEquals, expSnapSetup, Commentf(t.Kind()))
-	}
-
-	// verify snaps in the system state
-	var snapst snapstate.SnapState
-	err = snapstate.Get(s.state, "some-snap_instance", &snapst)
-	c.Assert(err, Equals, state.ErrNoState)
-
-	// the non-instance snap is still there
-	err = snapstate.Get(s.state, "some-snap", &snapst)
-	c.Assert(err, IsNil)
-}
-
-func (s *snapmgrTestSuite) TestParallelInstanceRemoveRunThroughOtherInstances(c *C) {
-	si := snap.SideInfo{
-		RealName: "some-snap",
-		Revision: snap.R(7),
-	}
-
-	s.state.Lock()
-	defer s.state.Unlock()
-
-	// pretend we have both a regular snap and a parallel instance
-	snapstate.Set(s.state, "some-snap_instance", &snapstate.SnapState{
-		Active:      true,
-		Sequence:    []*snap.SideInfo{&si},
-		Current:     si.Revision,
-		SnapType:    "app",
-		InstanceKey: "instance",
-	})
-	snapstate.Set(s.state, "some-snap_other", &snapstate.SnapState{
-		Active:      true,
-		Sequence:    []*snap.SideInfo{&si},
-		Current:     si.Revision,
-		SnapType:    "app",
-		InstanceKey: "other",
-	})
-
-	chg := s.state.NewChange("remove", "remove a snap")
-	ts, err := snapstate.Remove(s.state, "some-snap_instance", snap.R(0), nil)
-	c.Assert(err, IsNil)
-	chg.AddAll(ts)
-
-	s.state.Unlock()
-	s.settle(c)
-	s.state.Lock()
-
-	expected := fakeOps{
-		{
-			op:    "auto-disconnect:Doing",
-			name:  "some-snap_instance",
-			revno: snap.R(7),
-		},
-		{
-			op:   "remove-snap-aliases",
-			name: "some-snap_instance",
-		},
-		{
-			op:   "unlink-snap",
-			path: filepath.Join(dirs.SnapMountDir, "some-snap_instance/7"),
-		},
-		{
-			op:    "remove-profiles:Doing",
-			name:  "some-snap_instance",
-			revno: snap.R(7),
-		},
-		{
-			op:   "remove-snap-data",
-			path: filepath.Join(dirs.SnapMountDir, "some-snap_instance/7"),
-		},
-		{
-			op:   "remove-snap-common-data",
-			path: filepath.Join(dirs.SnapMountDir, "some-snap_instance/7"),
-		},
-		{
-			op:             "remove-snap-data-dir",
-			name:           "some-snap_instance",
-			path:           filepath.Join(dirs.SnapDataDir, "some-snap"),
-			otherInstances: true,
-		},
-		{
-			op:    "remove-snap-files",
-			path:  filepath.Join(dirs.SnapMountDir, "some-snap_instance/7"),
-			stype: "app",
-		},
-		{
-			op:   "discard-namespace",
-			name: "some-snap_instance",
-		},
-		{
-			op:             "remove-snap-dir",
-			name:           "some-snap_instance",
-			path:           filepath.Join(dirs.SnapMountDir, "some-snap"),
-			otherInstances: true,
-		},
-	}
-	// start with an easier-to-read error if this fails:
-	c.Check(len(s.fakeBackend.ops), Equals, len(expected))
-	c.Assert(s.fakeBackend.ops.Ops(), DeepEquals, expected.Ops())
-	c.Check(s.fakeBackend.ops, DeepEquals, expected)
-
-	// verify snaps in the system state
-	var snapst snapstate.SnapState
-	err = snapstate.Get(s.state, "some-snap_instance", &snapst)
-	c.Assert(err, Equals, state.ErrNoState)
-
-	// the other instance is still there
-	err = snapstate.Get(s.state, "some-snap_other", &snapst)
-	c.Assert(err, IsNil)
-}
-
-func (s *snapmgrTestSuite) TestRemoveWithManyRevisionsRunThrough(c *C) {
-	si3 := snap.SideInfo{
-		SnapID:   "some-snap-id",
-		RealName: "some-snap",
-		Revision: snap.R(3),
-	}
-
-	si5 := snap.SideInfo{
-		SnapID:   "some-snap-id",
-		RealName: "some-snap",
-		Revision: snap.R(5),
-	}
-
-	si7 := snap.SideInfo{
-		SnapID:   "some-snap-id",
-		RealName: "some-snap",
-		Revision: snap.R(7),
-	}
-
-	s.state.Lock()
-	defer s.state.Unlock()
-
-	snapstate.Set(s.state, "some-snap", &snapstate.SnapState{
-		Active:   true,
-		Sequence: []*snap.SideInfo{&si5, &si3, &si7},
-		Current:  si7.Revision,
-		SnapType: "app",
-	})
-
-	chg := s.state.NewChange("remove", "remove a snap")
-	ts, err := snapstate.Remove(s.state, "some-snap", snap.R(0), nil)
-	c.Assert(err, IsNil)
-	chg.AddAll(ts)
-
-	s.state.Unlock()
-	defer s.se.Stop()
-	s.settle(c)
-	s.state.Lock()
-
-	expected := fakeOps{
-		{
-			op:    "auto-disconnect:Doing",
-			name:  "some-snap",
-			revno: snap.R(7),
-		},
-		{
-			op:   "remove-snap-aliases",
-			name: "some-snap",
-		},
-		{
-			op:   "unlink-snap",
-			path: filepath.Join(dirs.SnapMountDir, "some-snap/7"),
-		},
-		{
-			op:    "remove-profiles:Doing",
-			name:  "some-snap",
-			revno: snap.R(7),
-		},
-		{
-			op:   "remove-snap-data",
-			path: filepath.Join(dirs.SnapMountDir, "some-snap/7"),
-		},
-		{
-			op:    "remove-snap-files",
-			path:  filepath.Join(dirs.SnapMountDir, "some-snap/7"),
-			stype: "app",
-		},
-		{
-			op:   "remove-snap-data",
-			path: filepath.Join(dirs.SnapMountDir, "some-snap/3"),
-		},
-		{
-			op:    "remove-snap-files",
-			path:  filepath.Join(dirs.SnapMountDir, "some-snap/3"),
-			stype: "app",
-		},
-		{
-			op:   "remove-snap-data",
-			path: filepath.Join(dirs.SnapMountDir, "some-snap/5"),
-		},
-		{
-			op:   "remove-snap-common-data",
-			path: filepath.Join(dirs.SnapMountDir, "some-snap/5"),
-		},
-		{
-			op:   "remove-snap-data-dir",
-			name: "some-snap",
-			path: filepath.Join(dirs.SnapDataDir, "some-snap"),
-		},
-		{
-			op:    "remove-snap-files",
-			path:  filepath.Join(dirs.SnapMountDir, "some-snap/5"),
-			stype: "app",
-		},
-		{
-			op:   "discard-namespace",
-			name: "some-snap",
-		},
-		{
-			op:   "remove-snap-dir",
-			name: "some-snap",
-			path: filepath.Join(dirs.SnapMountDir, "some-snap"),
-		},
-	}
-	// start with an easier-to-read error if this fails:
-	c.Assert(s.fakeBackend.ops.Ops(), DeepEquals, expected.Ops())
-	c.Assert(s.fakeBackend.ops, DeepEquals, expected)
-
-	// verify snapSetup info
-	tasks := ts.Tasks()
-	revnos := []snap.Revision{{N: 7}, {N: 3}, {N: 5}}
-	whichRevno := 0
-	for _, t := range tasks {
-		if t.Kind() == "run-hook" {
-			continue
-		}
-		if t.Kind() == "save-snapshot" {
-			continue
-		}
-		snapsup, err := snapstate.TaskSnapSetup(t)
-		c.Assert(err, IsNil)
-
-		var expSnapSetup *snapstate.SnapSetup
-		switch t.Kind() {
-		case "discard-conns":
-			expSnapSetup = &snapstate.SnapSetup{
-				SideInfo: &snap.SideInfo{
-					SnapID:   "some-snap-id",
-					RealName: "some-snap",
-				},
-			}
-		case "clear-snap", "discard-snap":
-			expSnapSetup = &snapstate.SnapSetup{
-				SideInfo: &snap.SideInfo{
-					SnapID:   "some-snap-id",
-					RealName: "some-snap",
-					Revision: revnos[whichRevno],
-				},
-			}
-		default:
-			expSnapSetup = &snapstate.SnapSetup{
-				SideInfo: &snap.SideInfo{
-					SnapID:   "some-snap-id",
-					RealName: "some-snap",
-					Revision: snap.R(7),
-				},
-				Type:      snap.TypeApp,
-				PlugsOnly: true,
-			}
-
-		}
-
-		c.Check(snapsup, DeepEquals, expSnapSetup, Commentf(t.Kind()))
-
-		if t.Kind() == "discard-snap" {
-			whichRevno++
-		}
-	}
-
-	// verify snaps in the system state
-	var snapst snapstate.SnapState
-	err = snapstate.Get(s.state, "some-snap", &snapst)
-	c.Assert(err, Equals, state.ErrNoState)
-}
-
-func (s *snapmgrTestSuite) TestRemoveOneRevisionRunThrough(c *C) {
-	si3 := snap.SideInfo{
-		RealName: "some-snap",
-		Revision: snap.R(3),
-	}
-
-	si5 := snap.SideInfo{
-		RealName: "some-snap",
-		Revision: snap.R(5),
-	}
-
-	si7 := snap.SideInfo{
-		RealName: "some-snap",
-		Revision: snap.R(7),
-	}
-
-	s.state.Lock()
-	defer s.state.Unlock()
-
-	snapstate.Set(s.state, "some-snap", &snapstate.SnapState{
-		Active:   true,
-		Sequence: []*snap.SideInfo{&si5, &si3, &si7},
-		Current:  si7.Revision,
-		SnapType: "app",
-	})
-
-	chg := s.state.NewChange("remove", "remove a snap")
-	ts, err := snapstate.Remove(s.state, "some-snap", snap.R(3), nil)
-	c.Assert(err, IsNil)
-	chg.AddAll(ts)
-
-	s.state.Unlock()
-	defer s.se.Stop()
-	s.settle(c)
-	s.state.Lock()
-
-	c.Check(len(s.fakeBackend.ops), Equals, 2)
-	expected := fakeOps{
-		{
-			op:   "remove-snap-data",
-			path: filepath.Join(dirs.SnapMountDir, "some-snap/3"),
-		},
-		{
-			op:    "remove-snap-files",
-			path:  filepath.Join(dirs.SnapMountDir, "some-snap/3"),
-			stype: "app",
-		},
-	}
-	// start with an easier-to-read error if this fails:
-	c.Assert(s.fakeBackend.ops.Ops(), DeepEquals, expected.Ops())
-	c.Assert(s.fakeBackend.ops, DeepEquals, expected)
-
-	// verify snapSetup info
-	tasks := ts.Tasks()
-	for _, t := range tasks {
-		if t.Kind() == "save-snapshot" {
-			continue
-		}
-		snapsup, err := snapstate.TaskSnapSetup(t)
-		c.Assert(err, IsNil)
-
-		expSnapSetup := &snapstate.SnapSetup{
-			SideInfo: &snap.SideInfo{
-				RealName: "some-snap",
-				Revision: snap.R(3),
-			},
-		}
-
-		c.Check(snapsup, DeepEquals, expSnapSetup, Commentf(t.Kind()))
-	}
-
-	// verify snaps in the system state
-	var snapst snapstate.SnapState
-	err = snapstate.Get(s.state, "some-snap", &snapst)
-	c.Assert(err, IsNil)
-	c.Check(snapst.Sequence, HasLen, 2)
-}
-
-func (s *snapmgrTestSuite) TestRemoveLastRevisionRunThrough(c *C) {
-	si := snap.SideInfo{
-		RealName: "some-snap",
-		Revision: snap.R(2),
-	}
-
-	s.state.Lock()
-	defer s.state.Unlock()
-
-	snapstate.Set(s.state, "some-snap", &snapstate.SnapState{
-		Active:   false,
-		Sequence: []*snap.SideInfo{&si},
-		Current:  si.Revision,
-		SnapType: "app",
-	})
-
-	chg := s.state.NewChange("remove", "remove a snap")
-	ts, err := snapstate.Remove(s.state, "some-snap", snap.R(2), nil)
-	c.Assert(err, IsNil)
-	chg.AddAll(ts)
-
-	s.state.Unlock()
-	defer s.se.Stop()
-	s.settle(c)
-	s.state.Lock()
-
-	c.Check(len(s.fakeBackend.ops), Equals, 7)
-	expected := fakeOps{
-		{
-			op:    "auto-disconnect:Doing",
-			name:  "some-snap",
-			revno: snap.R(2),
-		},
-		{
-			op:   "remove-snap-data",
-			path: filepath.Join(dirs.SnapMountDir, "some-snap/2"),
-		},
-		{
-			op:   "remove-snap-common-data",
-			path: filepath.Join(dirs.SnapMountDir, "some-snap/2"),
-		},
-		{
-			op:   "remove-snap-data-dir",
-			name: "some-snap",
-			path: filepath.Join(dirs.SnapDataDir, "some-snap"),
-		},
-		{
-			op:    "remove-snap-files",
-			path:  filepath.Join(dirs.SnapMountDir, "some-snap/2"),
-			stype: "app",
-		},
-		{
-			op:   "discard-namespace",
-			name: "some-snap",
-		},
-		{
-			op:   "remove-snap-dir",
-			name: "some-snap",
-			path: filepath.Join(dirs.SnapMountDir, "some-snap"),
-		},
-	}
-	// start with an easier-to-read error if this fails:
-	c.Assert(s.fakeBackend.ops.Ops(), DeepEquals, expected.Ops())
-	c.Assert(s.fakeBackend.ops, DeepEquals, expected)
-
-	// verify snapSetup info
-	tasks := ts.Tasks()
-	for _, t := range tasks {
-		if t.Kind() == "run-hook" {
-			continue
-		}
-		if t.Kind() == "save-snapshot" {
-			continue
-		}
-		snapsup, err := snapstate.TaskSnapSetup(t)
-		c.Assert(err, IsNil)
-
-		expSnapSetup := &snapstate.SnapSetup{
-			SideInfo: &snap.SideInfo{
-				RealName: "some-snap",
-			},
-		}
-		if t.Kind() != "discard-conns" {
-			expSnapSetup.SideInfo.Revision = snap.R(2)
-		}
-		if t.Kind() == "auto-disconnect" {
-			expSnapSetup.PlugsOnly = true
-			expSnapSetup.Type = "app"
-		}
-
-		c.Check(snapsup, DeepEquals, expSnapSetup, Commentf(t.Kind()))
-	}
-
-	// verify snaps in the system state
-	var snapst snapstate.SnapState
-	err = snapstate.Get(s.state, "some-snap", &snapst)
-	c.Assert(err, Equals, state.ErrNoState)
-}
-
-func (s *snapmgrTestSuite) TestRemoveCurrentActiveRevisionRefused(c *C) {
-	si := snap.SideInfo{
-		RealName: "some-snap",
-		Revision: snap.R(2),
-	}
-
-	s.state.Lock()
-	defer s.state.Unlock()
-
-	snapstate.Set(s.state, "some-snap", &snapstate.SnapState{
-		Active:   true,
-		Sequence: []*snap.SideInfo{&si},
-		Current:  si.Revision,
-		SnapType: "app",
-	})
-
-	_, err := snapstate.Remove(s.state, "some-snap", snap.R(2), nil)
-
-	c.Check(err, ErrorMatches, `cannot remove active revision 2 of snap "some-snap"`)
-}
-
-func (s *snapmgrTestSuite) TestRemoveCurrentRevisionOfSeveralRefused(c *C) {
-	si := snap.SideInfo{
-		RealName: "some-snap",
-		Revision: snap.R(2),
-	}
-
-	s.state.Lock()
-	defer s.state.Unlock()
-
-	snapstate.Set(s.state, "some-snap", &snapstate.SnapState{
-		Active:   true,
-		Sequence: []*snap.SideInfo{&si, &si},
-		Current:  si.Revision,
-		SnapType: "app",
-	})
-
-	_, err := snapstate.Remove(s.state, "some-snap", snap.R(2), nil)
-	c.Assert(err, NotNil)
-	c.Check(err.Error(), Equals, `cannot remove active revision 2 of snap "some-snap" (revert first?)`)
-}
-
-func (s *snapmgrTestSuite) TestRemoveMissingRevisionRefused(c *C) {
-	si := snap.SideInfo{
-		RealName: "some-snap",
-		Revision: snap.R(2),
-	}
-
-	s.state.Lock()
-	defer s.state.Unlock()
-
-	snapstate.Set(s.state, "some-snap", &snapstate.SnapState{
-		Active:   true,
-		Sequence: []*snap.SideInfo{&si},
-		Current:  si.Revision,
-		SnapType: "app",
-	})
-
-	_, err := snapstate.Remove(s.state, "some-snap", snap.R(1), nil)
-
-	c.Check(err, ErrorMatches, `revision 1 of snap "some-snap" is not installed`)
-}
-
-func (s *snapmgrTestSuite) TestRemoveRefused(c *C) {
-	si := snap.SideInfo{
-		RealName: "brand-gadget",
-		Revision: snap.R(7),
-	}
-
-	s.state.Lock()
-	defer s.state.Unlock()
-
-	snapstate.Set(s.state, "brand-gadget", &snapstate.SnapState{
-		Active:   true,
-		Sequence: []*snap.SideInfo{&si},
-		Current:  si.Revision,
-		SnapType: "gadget",
-	})
-
-	_, err := snapstate.Remove(s.state, "brand-gadget", snap.R(0), nil)
-
-	c.Check(err, ErrorMatches, `snap "brand-gadget" is not removable: snap is used by the model`)
-}
-
-func (s *snapmgrTestSuite) TestRemoveRefusedLastRevision(c *C) {
-	si := snap.SideInfo{
-		RealName: "brand-gadget",
-		Revision: snap.R(7),
-	}
-
-	s.state.Lock()
-	defer s.state.Unlock()
-
-	snapstate.Set(s.state, "brand-gadget", &snapstate.SnapState{
-		Active:   false,
-		Sequence: []*snap.SideInfo{&si},
-		Current:  si.Revision,
-		SnapType: "gadget",
-	})
-
-	_, err := snapstate.Remove(s.state, "brand-gadget", snap.R(7), nil)
-
-	c.Check(err, ErrorMatches, `snap "brand-gadget" is not removable: snap is used by the model`)
-}
-
-func (s *snapmgrTestSuite) TestRemoveDeletesConfigOnLastRevision(c *C) {
-	si := snap.SideInfo{
-		RealName: "some-snap",
-		Revision: snap.R(7),
-	}
-
-	s.state.Lock()
-	defer s.state.Unlock()
-
-	snapstate.Set(s.state, "some-snap", &snapstate.SnapState{
-		Active:   true,
-		Sequence: []*snap.SideInfo{&si},
-		Current:  si.Revision,
-		SnapType: "app",
-	})
-
-	snapstate.Set(s.state, "another-snap", &snapstate.SnapState{
-		Active:   true,
-		Sequence: []*snap.SideInfo{&si},
-		Current:  si.Revision,
-		SnapType: "app",
-	})
-
-	tr := config.NewTransaction(s.state)
-	tr.Set("some-snap", "foo", "bar")
-	tr.Commit()
-
-	// a config for some other snap to verify its not accidentally destroyed
-	tr = config.NewTransaction(s.state)
-	tr.Set("another-snap", "bar", "baz")
-	tr.Commit()
-
-	var res string
-	tr = config.NewTransaction(s.state)
-	c.Assert(tr.Get("some-snap", "foo", &res), IsNil)
-	c.Assert(tr.Get("another-snap", "bar", &res), IsNil)
-
-	chg := s.state.NewChange("remove", "remove a snap")
-	ts, err := snapstate.Remove(s.state, "some-snap", snap.R(0), nil)
-	c.Assert(err, IsNil)
-	chg.AddAll(ts)
-
-	s.state.Unlock()
-	defer s.se.Stop()
-	s.settle(c)
-	s.state.Lock()
-
-	// verify snaps in the system state
-	var snapst snapstate.SnapState
-	err = snapstate.Get(s.state, "some-snap", &snapst)
-	c.Assert(err, Equals, state.ErrNoState)
-
-	tr = config.NewTransaction(s.state)
-	err = tr.Get("some-snap", "foo", &res)
-	c.Assert(err, NotNil)
-	c.Assert(err, ErrorMatches, `snap "some-snap" has no "foo" configuration option`)
-
-	// and another snap has its config intact
-	c.Assert(tr.Get("another-snap", "bar", &res), IsNil)
-	c.Assert(res, Equals, "baz")
-}
-
-func (s *snapmgrTestSuite) TestRemoveDoesntDeleteConfigIfNotLastRevision(c *C) {
-	si1 := snap.SideInfo{
-		RealName: "some-snap",
-		Revision: snap.R(7),
-	}
-	si2 := snap.SideInfo{
-		RealName: "some-snap",
-		Revision: snap.R(8),
-	}
-
-	s.state.Lock()
-	defer s.state.Unlock()
-
-	snapstate.Set(s.state, "some-snap", &snapstate.SnapState{
-		Active:   true,
-		Sequence: []*snap.SideInfo{&si1, &si2},
-		Current:  si2.Revision,
-		SnapType: "app",
-	})
-
-	tr := config.NewTransaction(s.state)
-	tr.Set("some-snap", "foo", "bar")
-	tr.Commit()
-
-	var res string
-	tr = config.NewTransaction(s.state)
-	c.Assert(tr.Get("some-snap", "foo", &res), IsNil)
-
-	chg := s.state.NewChange("remove", "remove a snap")
-	ts, err := snapstate.Remove(s.state, "some-snap", si1.Revision, nil)
-	c.Assert(err, IsNil)
-	chg.AddAll(ts)
-
-	s.state.Unlock()
-	defer s.se.Stop()
-	s.settle(c)
-	s.state.Lock()
-
-	// verify snaps in the system state
-	var snapst snapstate.SnapState
-	err = snapstate.Get(s.state, "some-snap", &snapst)
-	c.Assert(err, IsNil)
-
-	tr = config.NewTransaction(s.state)
-	c.Assert(tr.Get("some-snap", "foo", &res), IsNil)
-	c.Assert(res, Equals, "bar")
 }
 
 func (s *snapmgrTestSuite) TestRevertRestoresConfigSnapshot(c *C) {
@@ -5267,165 +4146,6 @@ func (s *snapmgrTestSuite) TestSnapStateLocalRevision(c *C) {
 		Current:  si7.Revision,
 	}
 	c.Assert(snapst.LocalRevision().Unset(), Equals, true)
-}
-
-func (s *snapmgrTestSuite) TestRemoveMany(c *C) {
-	s.state.Lock()
-	defer s.state.Unlock()
-
-	snapstate.Set(s.state, "one", &snapstate.SnapState{
-		Active: true,
-		Sequence: []*snap.SideInfo{
-			{RealName: "one", SnapID: "one-id", Revision: snap.R(1)},
-		},
-		Current: snap.R(1),
-	})
-	snapstate.Set(s.state, "two", &snapstate.SnapState{
-		Active: true,
-		Sequence: []*snap.SideInfo{
-			{RealName: "two", SnapID: "two-id", Revision: snap.R(1)},
-		},
-		Current: snap.R(1),
-	})
-
-	removed, tts, err := snapstate.RemoveMany(s.state, []string{"one", "two"})
-	c.Assert(err, IsNil)
-	c.Assert(tts, HasLen, 2)
-	c.Check(removed, DeepEquals, []string{"one", "two"})
-
-	c.Assert(s.state.TaskCount(), Equals, 8*2)
-	for i, ts := range tts {
-		c.Assert(taskKinds(ts.Tasks()), DeepEquals, []string{
-			"stop-snap-services",
-			"run-hook[remove]",
-			"auto-disconnect",
-			"remove-aliases",
-			"unlink-snap",
-			"remove-profiles",
-			"clear-snap",
-			"discard-snap",
-		})
-		verifyStopReason(c, ts, "remove")
-		// check that tasksets are in separate lanes
-		for _, t := range ts.Tasks() {
-			c.Assert(t.Lanes(), DeepEquals, []int{i + 1})
-		}
-
-	}
-}
-
-func (s *snapmgrTestSuite) testRemoveManyDiskSpaceCheck(c *C, featureFlag, automaticSnapshot, freeSpaceCheckFail bool) error {
-	s.state.Lock()
-	defer s.state.Unlock()
-
-	var checkFreeSpaceCall, snapshotSizeCall int
-
-	// restored by TearDownTest
-	snapstate.EstimateSnapshotSize = func(st *state.State, instanceName string, users []string) (uint64, error) {
-		snapshotSizeCall++
-		// expect two snapshot size estimations
-		switch instanceName {
-		case "one":
-			return 10, nil
-		case "two":
-			return 20, nil
-		default:
-			c.Fatalf("unexpected snap: %s", instanceName)
-		}
-		return 1, nil
-	}
-
-	restore := snapstate.MockOsutilCheckFreeSpace(func(path string, required uint64) error {
-		checkFreeSpaceCall++
-		// required size is the sum of snapshot sizes of test snaps
-		c.Check(required, Equals, snapstate.SafetyMarginDiskSpace(30))
-		if freeSpaceCheckFail {
-			return &osutil.NotEnoughDiskSpaceError{}
-		}
-		return nil
-	})
-	defer restore()
-
-	var automaticSnapshotCalled bool
-	snapstate.AutomaticSnapshot = func(st *state.State, instanceName string) (ts *state.TaskSet, err error) {
-		automaticSnapshotCalled = true
-		if automaticSnapshot {
-			t := s.state.NewTask("foo", "")
-			ts = state.NewTaskSet(t)
-			return ts, nil
-		}
-		// ErrNothingToDo is returned if automatic snapshots are disabled
-		return nil, snapstate.ErrNothingToDo
-	}
-
-	tr := config.NewTransaction(s.state)
-	tr.Set("core", "experimental.check-disk-space-remove", featureFlag)
-	tr.Commit()
-
-	snapstate.Set(s.state, "one", &snapstate.SnapState{
-		Active:   true,
-		SnapType: "app",
-		Sequence: []*snap.SideInfo{
-			{RealName: "one", SnapID: "one-id", Revision: snap.R(1)},
-		},
-		Current: snap.R(1),
-	})
-	snapstate.Set(s.state, "two", &snapstate.SnapState{
-		Active:   true,
-		SnapType: "app",
-		Sequence: []*snap.SideInfo{
-			{RealName: "two", SnapID: "two-id", Revision: snap.R(1)},
-		},
-		Current: snap.R(1),
-	})
-
-	_, _, err := snapstate.RemoveMany(s.state, []string{"one", "two"})
-	if featureFlag && automaticSnapshot {
-		c.Check(snapshotSizeCall, Equals, 2)
-		c.Check(checkFreeSpaceCall, Equals, 1)
-	} else {
-		c.Check(checkFreeSpaceCall, Equals, 0)
-		c.Check(snapshotSizeCall, Equals, 0)
-	}
-	c.Check(automaticSnapshotCalled, Equals, true)
-
-	return err
-}
-
-func (s *snapmgrTestSuite) TestRemoveManyDiskSpaceError(c *C) {
-	featureFlag := true
-	automaticSnapshot := true
-	freeSpaceCheckFail := true
-	err := s.testRemoveManyDiskSpaceCheck(c, featureFlag, automaticSnapshot, freeSpaceCheckFail)
-
-	diskSpaceErr := err.(*snapstate.InsufficientSpaceError)
-	c.Check(diskSpaceErr.Path, Equals, filepath.Join(dirs.GlobalRootDir, "/var/lib/snapd"))
-	c.Check(diskSpaceErr.Snaps, DeepEquals, []string{"one", "two"})
-	c.Check(diskSpaceErr.ChangeKind, Equals, "remove")
-}
-
-func (s *snapmgrTestSuite) TestRemoveManyDiskSpaceCheckDisabled(c *C) {
-	featureFlag := false
-	automaticSnapshot := true
-	freeSpaceCheckFail := true
-	err := s.testRemoveManyDiskSpaceCheck(c, featureFlag, automaticSnapshot, freeSpaceCheckFail)
-	c.Assert(err, IsNil)
-}
-
-func (s *snapmgrTestSuite) TestRemoveManyDiskSpaceSnapshotDisabled(c *C) {
-	featureFlag := true
-	automaticSnapshot := false
-	freeSpaceCheckFail := true
-	err := s.testRemoveManyDiskSpaceCheck(c, featureFlag, automaticSnapshot, freeSpaceCheckFail)
-	c.Assert(err, IsNil)
-}
-
-func (s *snapmgrTestSuite) TestRemoveManyDiskSpaceCheckPasses(c *C) {
-	featureFlag := true
-	automaticSnapshot := true
-	freeSpaceCheckFail := false
-	err := s.testRemoveManyDiskSpaceCheck(c, featureFlag, automaticSnapshot, freeSpaceCheckFail)
-	c.Check(err, IsNil)
 }
 
 func tasksWithKind(ts *state.TaskSet, kind string) []*state.Task {
